Sprindiks valmis

+++ AnaluusiKomponent/IAnaluus.vb
@@ -1,7 +1,7 @@
 ﻿Public Interface IAnaluus
     Function paevaneKcal(kcalHommik, kcalLouna, kcalOhtu, kcalVahepala) As Integer
-    Function KclParingAndmebaasist(kasutajaId As Integer, kuupaev As Integer, toidukord As Integer) As Integer()
-    Function ToidukordKokku(ByRef KcalLoend() As Integer) As Integer
+    Function KclParingAndmebaasist(kasutajaId As Integer, kuupaev As Integer, toidukord As Integer) As Double()
+    Function ToidukordKokku(ByRef KcalLoend() As Double) As Double
     Function KuupaevIntegeriks(ByVal sisendKuupaev As DateTime) As Integer
     Function IntegerKuupaevaks(ByVal sisendInteger As Integer) As DateTime
 End Interface
